+++ changelog.md
@@ -5,7 +5,8 @@ ChangeLog
 * Regenerate index at bookmark removal
 * Add context menu for non-system-folders
 * Removed unused session variable
-* clientupdates
+* Changed Client Updates
+* Changed Import to transaction
 
 
 1.2.17 (2020-12-12)